Fixed threading!
The loading process is threaded. I needs to load pages as well as new
posts still, but it works. The issue was that you can't call an Async
Task from an Async Task. It can only be called from the UI thread.
However, you CAN call a Runnable thread from an Async task. So now all
the downloading is done via implemented Runnable threads instead of
Async tasks, and it works inside and outside of the main UI thread.